api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.6.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pandas>=2.0.0
httpx>=0.25.0
pydantic>=2.10.0
python-multipart>=0.0.6
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
langchain-nvidia-ai-endpoints>=1.1.0
cachetools>=7.0.0
